gwa_studio_llms/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai # Para el cliente de Gemini